+++ .github/workflows/eBPF_proc_image.yml
@@ -49,16 +49,10 @@ jobs:
           cd eBPF_Supermarket/CPU_Subsystem/eBPF_proc_image/tools
           make keytime_image
           sudo ./keytime_image -t 1
-
+      
       - name: Run proc_image
         run: |
           cd eBPF_Supermarket/CPU_Subsystem/eBPF_proc_image
-          make proc_image
-          sudo ./proc_image -t 1
-      
-      - name: Run new_proc_image
-        run: |
-          cd eBPF_Supermarket/CPU_Subsystem/eBPF_proc_image/developing
           make proc_image
           sudo ./proc_image -a -p 1 -t 1
